client/requirements.txt
streamlit==1.41.1
python-dotenv
groq
pyyaml
requests